Implemented jpa for user dao and book dao

+++ src/main/java/com/belhard/bookstore/dao/BookDao.java
@@ -20,6 +20,6 @@ public interface BookDao {
 
     boolean deleteBook(Long id);
 
-    int countAllBooks();
+    Long countAllBooks();
 }
